added a non-game
this was also part of the AI assignment. However this isn't really a game. More a pure school assignment. This because you don't play in anyway. The test was to make the hat able to play the game on it's own by collecting the right item and dodging the bombs